Refactor simple payment module
The _make_json_request_data was an unnecessary abstraction. It's one line of
code, not used anywhere else, makes the reader an unncessary jump. The variable
has a meaningful name, so the meaning and the purpose is clear.
Might need to be reintroduced later when multiple functions should call it.

hmac.digest is equivalent with making an hmac digest and calling the .digest()
method on it, but this way it's more succint.

"data" is not meaningful in a variable name.

BASE_URL was a bad variable name