Email verification
Email is verified with a code, which is used as a salt value for the Django
Signer, so every email can be verified only with the appropriate salt. This way,
we don't need to put the code in the email link, but we can later if we decide
that would be easier for the users.