Handle email verification token
Generate a token for an EmailVerification and get the same Emailverification
by that token.

The token is first signed with Django Signer:
https://docs.djangoproject.com/en/3.0/topics/signing/

using EmailVerification.code as salt value, then encrypted with the
EMAIL_VERIFICATION_KEY, which is a symmetric Fernet encryption key:
https://cryptography.io/en/latest/fernet/

When getting the key from the token, it is decrypted, then we split the
EmailVerification.uuid without checking the signature, because we don't have a
salt yet (might be suspectible for timing attacks?), then we find the relevant
EmailVerification by UUID, then verify the signature and set the verified_at
date. The correct order is sign-than-encrypt:
https://crypto.stackexchange.com/questions/5458/should-we-sign-then-encrypt-or-encrypt-then-sign

There is a test for the full workflow.