Validate total_price on /pay-appointment/ endpoint
It is just a sanity check, total_price will NOT be calculated based on the sent
request at all, it's just compared with the shown price, so we make sure the
user don't get an amount they didn't see on the frontend.

Also detecting suspicious operations (e.g. somebody tried to change
the total_price).