Make assertions for the total expected Payments
There is an assertion at the end of every test case for /api/get-price/ and
/api/pay-appointment/ to make sure the expected number of Payment objects has
been saved in the database.
In case of /get-price/ this is always 0, with no new Payment records.
In case of /pay-appointment/ this depends on a lot of variables, but the number
of seats should always match with the number of Payment objects on successful
api calls.