Enabled SessionAuthentication and reworked Seat resource redirect logic
When logged in via token authentication (using the Authorization header), we
return the response as it is. When the user is logged in via
SessionAuthentication, we first check if they are browsing the API with the HTML
API browser, with the ?format=api URL param set, when this is the case, we
render the response the same way. Any other case, we redirect to the Seat admin
page.
The request.user is always set, but request.auth is only set when logged in via
TokenAuthentication backend.