Implement Django Persmission handling on /staff-api/
Whatever Groups the user is in, the /staff-api/ will allow or deny access based
on the model Permissions belonging to that Group for the relevant resources.
Disallow setting permissions on the user individually and deleting groups, so
nobody can mess up the permission settings, the Mobile app will use these.