Normalize licence plate on the backend instead of the frontend
In this project the flow only allows adding licence plates as an update to the
Appointments, but not during the creation.

It doesn't make any sense to do any licence plate normalization on the frontend
size, since it must be normalized on the backend anyway, therefore it got
removed from the frontend.

Look at the tests to understand the expected behaviour of licence plate
normalizer.

If the licence plate normalization will happen at multiple places, consider
adding a custom method the Appointment model to do that, for data integrity.