Added QRCode model with autogenerating code
We needed a separate model, because we need a monotonic number for the manual
code, which is described in the Project's wiki:
https://gitlab.com/rollet/project-noe/-/wikis/Seat-QR-codes

The basic idea is that we first save the object to get the id from the primary
key, and that will be the last four digits, the two other parts are changing by
location and by day.

The code field keeps the code which the URL is generated from, and it's unique.
There are tests which test this code generation behavior.