Send qrcode for every seat
It is not enough to send only one email for every Seat even if the email address
is the same, because the QR codes will be different, so in that case, we need to
send as many emails as many Seats are in the Appointment.

the email.send_qrcode now only expects a seat (and a seat_count
for information), because every necessary information is available on the seat.