Fix Appointment and Seat permissions
Added Appointment permission checks to /get-price/ and /pay-appointment/
endpoints.

has_object_permission is a method on the Permission class, NOT on the view, but
in order to customize for every view (as appointment is in a different attribute
on different views) we need to call a method on the view from the permission
class. (The other solution would have been to pass a lambda function for the
permission class, but I found this cleaner.)