Refact check_same_appointment
Mos of the time, it was the same object, checked obj agains request.auth.
We can simplify the common case make that the default, and the view should only
define a custom method for accessing the relevant appointment if it's more
complicated than that.

There is an assert for checking the get_appointment fuction returning an
Appointment instance, because we don't use mypy yet.