Fix billing never sent, refactor validate_paid_at and handle_paid_at
When calling super().update() from the serializer, it will update the instance,
so the payment.paid_at will be changed, and we are not comparing the values we
intended to compare (the original field in the database and the new submitted
value). We don't actually need the payment instance, so pass in the original
value of paid_at directly, so we can avoid problems around mutated objects.